+++ BillingModule/billingModule/src/billingmodule/DataFreeUnitsCalc.java
@@ -24,8 +24,8 @@ public class DataFreeUnitsCalc {
         ProfileService profileDataDetails;
         Boolean state = false;
             
-        int updatedValue = 0;
-        int consumedData = 0;
+        Float updatedValue = 0f;
+        Float consumedData = 0f;
         Float costOfService = 0f;
         Float TotalUDRsCost = 0f;
         
@@ -65,7 +65,7 @@ public class DataFreeUnitsCalc {
                                 }else{
                                       
                                         cProfileupdateFU = new CustomerProfile(
-                                                    udr.getDialA(),udr.getProfileID(),udr.getServiceID(),0);  
+                                                    udr.getDialA(),udr.getProfileID(),udr.getServiceID(),0f);  
                                         state=db.UpdateCustomerFUs(cProfileupdateFU,"nothing");
                                         consumedData = abs(updatedValue);                                          
                                         costOfService = (consumedData * profileDataDetails.getFeeSameOperator())
