+++ BillingModule/billingModule/src/billingmodule/VoiceFreeUnitsCalc.java
@@ -30,8 +30,8 @@ public class VoiceFreeUnitsCalc {
         ProfileService profileVoiceDetails;
         Boolean state = false ;
         
-        int updatedValue =0;
-        int consumedData = 0;
+        Float updatedValue =0f;
+        Float consumedData = 0f;
         Float costOfService = 0f;
         Float TotalUDRsCost = 0f;
         Date start_date;
@@ -75,7 +75,7 @@ public class VoiceFreeUnitsCalc {
                             }else{ // cost of (-50) ---> profile_service
 
                                     cProfileupdateFU = new CustomerProfile(
-                                                udr.getDialA(),udr.getProfileID(),udr.getServiceID(),0);
+                                                udr.getDialA(),udr.getProfileID(),udr.getServiceID(),0f);
                                     state=db.UpdateCustomerFUs(cProfileupdateFU,NetConnection.onNet);
                                     consumedData = abs(updatedValue);  
                                     costOfService = (consumedData * profileVoiceDetails.getFeeSameOperator())
@@ -115,7 +115,7 @@ public class VoiceFreeUnitsCalc {
               
                             }else{
                                     cProfileupdateFU = new CustomerProfile(
-                                                udr.getDialA(),udr.getProfileID(),udr.getServiceID(),0);
+                                                udr.getDialA(),udr.getProfileID(),udr.getServiceID(),0f);
                                     state=db.UpdateCustomerFUs(cProfileupdateFU,NetConnection.crossNet);
                                     consumedData = abs(updatedValue);  
                                     costOfService = (consumedData * profileVoiceDetails.getFeeAnotherOperator())
